Fix wrong Hackage document URLs.
My previous logic for determining whether to add a Hackage document URL
was too loose and added links even for some internal code (mostly for
generated code).

Now reads out the list of packages that are shipped with GHC and adds
a Hackage link only when the span doesn't exist *and* it comes from
those core package.