Don't generate module tickets for implicit imports (Prelude)
They generate lots of stray Kythe edges because the target nodes don't
exist. Their anchors span the whole module declaration (e.g., module
Foo.Bar where), which is not useful.

And this may be a problem once
https://github.com/google/haskell-indexer/issues/127 is fixed because
the module declaration lines will then link to the Hackage document for
Prelude, which will be surprising to users.